提交人alexyouyi 挑战尝试闯关到Day11

+++ alexyouyi/day9_ScientificCalculator.sol
@@ -15,7 +15,7 @@ contract ScientificCalculator{
             result=(result + number / result)/2;
         }
 
-        return result;
+        return uint256(result);
     }
